Move facial recognition video into own component
